KE-13730 remove grafana public/test directory

+++ build/script_newten/download-grafana.sh
@@ -25,4 +25,7 @@ fi
 
 tar -zxvf build/grafana-${grafana_version}.linux-amd64.tar.gz -C build/ || { exit 1; }
 mv build/grafana-${grafana_version} build/grafana
-rm -rf build/grafana/conf/provisioning
+
+test -d 'build/grafana/conf/provisioning' && rm -rf build/grafana/conf/provisioning
+
+test -d 'build/grafana/public/test' && rm -rf build/grafana/public/test
